+++ modules/blackhole/cuda/blackhole_cuda.h
@@ -2,6 +2,6 @@
 #define __OPENSPACE_MODULE_BLACKHOLE___CUDA___H__
 #include <vector>
 
-void schwarzchild(float rs, float envmap_r, int num_paths, int num_steps, float u_0, float h, float* angle_out);
+void schwarzchild(float rs, float envmap_r, size_t num_paths, size_t num_steps, float u_0, float h, float* angle_out);
 
 #endif
